Refactor RepLine and modified getter and setter for Heading attribute.
