+++ requirements.txt
@@ -2,3 +2,4 @@ SQLAlchemy==1.3.10
 geoalchemy2
 psycopg2
 nose2==0.9.1
+pint==0.8
